requirements.txt
# Core dependencies for TrellisBMA
# DNA storage trace reconstruction using Trellis BMA algorithm

# Numerical computing
numpy==1.18.1
scipy==1.4.1

# Data manipulation and analysis
pandas==1.0.1

# Visualization
matplotlib==3.1.3
seaborn==0.10.0
matplotlib-label-lines

# Performance - JIT compilation (critical for algorithm performance)
numba==0.48.0

# Progress tracking
tqdm==4.42.1

# String/sequence distance calculations
python-Levenshtein==0.12.0

# Jupyter notebook support (for interactive documentation)
jupyter>=1.0.0
ipython>=7.0.0
